refactor(cache): Utils to calculate soft and hard TTL
